Api Key Request Guard
