+++ src/main.rs
@@ -12,12 +12,11 @@ mod user_control;
 mod utils;
 mod version_check;
 mod fairings;
+mod request_guard;
 
 use dotenv::dotenv;
 
-use rocket::http::Status;
-use rocket::log::private::info;
-use rocket::request::{FromRequest, Outcome, Request};
+use rocket::request::Request;
 
 
 use oracle::pool::PoolBuilder;
@@ -44,8 +43,6 @@ use routes::logs::delete_user_logs;
 use routes::logs::get_all_logs;
 use routes::version_check::route_version_check;
 // use crate::routes::user_control::edit_user;
-
-use signing::validate_token;
 
 use crate::fairings::log::Logger;
 use crate::fairings::cors::CORS;
@@ -130,39 +127,6 @@ fn rocket() -> _ {
         
 }
 
-// Start Request Guard Functions
-#[derive(Debug, Clone)]
-pub struct ApiKey<'r>(&'r str);
-
-#[rocket::async_trait]
-impl<'r> FromRequest<'r> for ApiKey<'r> {
-    type Error = String;
-
-    async fn from_request(req: &'r Request<'_>) -> Outcome<Self, Self::Error> {
-        // Returns true if `key` is a valid JWT Token.
-
-        match req.headers().get_one("Authorization") {
-            None => {
-                error!("No Authentication header found");
-                Outcome::Error((
-                    Status::Unauthorized,
-                    "Please include an Authentication header".to_string(),
-                ))
-            }
-            Some(key) if validate_token(key) => {
-                info!("Valid Token Found");
-                Outcome::Success(ApiKey(key))
-            }
-            Some(_) => {
-                error!("Invalid Token Found");
-                Outcome::Error((
-                    Status::Unauthorized,
-                    "Please include a valid Authentication header".to_string(),
-                ))
-            }
-        }
-    }
-}
 // Route catchers
 
 #[catch(400)]
