+++ src/routes/user_control.rs
@@ -2,7 +2,7 @@ use crate::signing::decode_token_data;
 use crate::user_control::structs::*;
 use crate::user_control::*;
 use crate::utils::structs::APIErrors;
-use crate::ApiKey;
+use crate::request_guard::api_key::ApiKey;
 
 use oracle::pool::Pool;
